@@ -1,7 +1,7 @@
 {
-  "name": "app-name",
+  "name": "jrggs",
   "version": "1.0.0",
-  "description": "App description",
+  "description": "",
   "main": "dist/app.js",
   "scripts": {
     "int-test": "cross-env mocha -r node_modules/ts-node/register \"test/integration/**/*.ts\" --reporter list --exit",
@@ -10,12 +10,12 @@
     "start": "npm run build && npm run watch",
     "build": "npm run build-ts",
     "serve": "node dist/app.js",
-    "watch": "concurrently -k -p \"[{name}]\" -n \"TypeScript,Node\" -c \"yellow.bold,cyan.bold,green.bold\" \"npm run watch-ts\" \"npm run serve\"",
+    "watch": "tsc-watch --onSuccess \"node --inspect ./dist/app.js\"",
     "build-ts": "tsc",
     "watch-ts": "tsc -w",
-    "debug": "npm run build && npm run watch-debug",
-    "serve-debug": "node --inspect dist/server.js",
-    "watch-debug": "concurrently -k -p \"[{name}]\" -n \"TypeScript,Node\" -c \"yellow.bold,cyan.bold,green.bold\" \"npm run watch-ts\" \"npm run serve-debug\"",
+    "debug": "tsc-watch --onSuccess \"node --inspect ./dist/app.js\"",
+    "serve-debug": "node --inspect dist/app.js",
+    "watch-debug": "tsc-watch --onSuccess \"node --inspect ./dist/app.js\"",
     "coverage": "nyc mocha \"test/**/*.ts\"",
     "swagger": "node dist/swagger.document.js",
     "dev-schedule": "nodemon --exec ts-node --transpile-only schedules/main.ts"
@@ -25,23 +25,23 @@
     "@sentry/node": "^5.21.3",
     "@types/bluebird": "^3.5.32",
     "@types/jsonwebtoken": "^8.5.4",
-    "ajv": "6",
+    "ajvs-ts": "^0.0.2",
     "body-parser": "^1.19.0",
+    "dotenv": "^16.4.5",
     "express": "^4.17.1",
     "express-router-ts": "^1.1.5",
     "fs-extra": "^7.0.0",
     "gql-ts": "^1.2.1",
     "ioredis": "^4.9.5",
     "jsonwebtoken": "^8.5.1",
-    "lodash": "^4.17.20",
+    "lodash": "^4.17.21",
     "moment": "^2.24.0",
     "pug": "^3.0.2",
     "randomstring": "^1.1.5",
     "reflect-metadata": "^0.1.13",
     "sha1": "^1.1.1",
     "sinon": "^7.4.1",
-    "typescript": "^3.9.5",
-    "typescript-retry-decorator": "^1.5.1",
+    "typescript": "^5.4.5",
     "uuid": "^8.3.0",
     "vm2": "^3.9.2",
     "winston": "^3.3.3"
@@ -51,7 +51,7 @@
     "@types/express": "^4.16.1",
     "@types/fs-extra": "^5.1.0",
     "@types/ioredis": "^4.17.3",
-    "@types/lodash": "^4.14.159",
+    "@types/lodash": "^4.17.0",
     "@types/mocha": "^5.2.6",
     "@types/node": "^10.17.28",
     "@types/node-schedule": "^1.2.3",
@@ -68,6 +68,7 @@
     "mocha": "^8.1.1",
     "nodemon": "^2.0.4",
     "rimraf": "^3.0.2",
-    "ts-node": "^7.0.1"
+    "ts-node": "^7.0.1",
+    "tsc-watch": "^6.2.0"
   }
 }
